requirements.txt
cryptography>=41.0.0
apscheduler>=3.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
python-dateutil>=2.8.2
pyyaml>=6.0
paho-mqtt>=1.6.1
python-telegram-bot>=20.0
google-genai>=0.1.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
sounddevice>=0.4.6
numpy>=1.26.0
scipy>=1.11.0
playwright>=1.40.0
openai>=1.3.0
flask>=3.0.0
flask-cors>=4.0.0
